requirements.txt
aniso8601==9.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
DateTime==4.5
Flask==2.1.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
influxdb-client==1.30.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
msgpack==1.0.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
Rx==3.2.0
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
zope.interface==5.4.0
